0.0.7: Added arm64 slice to simulator build
