+++ Roam/Roam.xcframework/Info.plist
@@ -6,11 +6,12 @@
 	<array>
 		<dict>
 			<key>LibraryIdentifier</key>
-			<string>ios-i386_x86_64-simulator</string>
+			<string>ios-arm64_i386_x86_64-simulator</string>
 			<key>LibraryPath</key>
 			<string>Roam.framework</string>
 			<key>SupportedArchitectures</key>
 			<array>
+				<string>arm64</string>
 				<string>i386</string>
 				<string>x86_64</string>
 			</array>
