+++ Roam/Roam.xcframework/ios-arm64_i386_x86_64-simulator/Roam.framework/Modules/module.modulemap
@@ -1,0 +1,11 @@
+framework module Roam {
+  umbrella header "Roam.h"
+
+  export *
+  module * { export * }
+}
+
+module Roam.Swift {
+    header "Roam-Swift.h"
+    requires objc
+}
