--- Roam/Roam.xcframework/ios-i386_x86_64-simulator/Roam.framework/Modules/module.modulemap
@@ -1,11 +1,0 @@
-framework module Roam {
-  umbrella header "Roam.h"
-
-  export *
-  module * { export * }
-}
-
-module Roam.Swift {
-    header "Roam-Swift.h"
-    requires objc
-}
